move botcase
