+++ db.js
@@ -1,0 +1,15 @@
+var MongoClient = require('mongodb').MongoClient;
+var url = "mongodb://localhost:27017/";
+
+MongoClient.connect(url, function(err, db) {
+	if (err) throw err;
+	var dbo = db.db("digimon");
+	var myobj = { 	_id: userID,
+					user: user,
+				};
+	dbo.collection("player").insertOne(myobj, function(err, res) {
+		if (err) throw err;
+			console.log("Insert Succeed!");
+		db.close();
+	});
+});
